Refactor migration of old settings and cache into separate module
